@@ -2,88 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:solarsense/modules/map/controllers/map_controller.dart';
+import 'package:solarsense/modules/map/state/map_state.dart';
 import 'package:solarsense/shared/constants/constants.dart';
-import 'package:solarsense/shared/constants/utilities.dart';
 import 'package:solarsense/shared/widgets/loading_indicator.dart';
-import 'package:geolocator/geolocator.dart';
 
-class MapView extends StatefulWidget {
+class MapView extends GetView<MapController> {
   const MapView({Key? key}) : super(key: key);
-
-  @override
-  State<MapView> createState() => _MapViewState();
-}
-
-class _MapViewState extends State<MapView> {
-  static const LatLng default_location =
-      LatLng(33.67715534264639, 73.06801369657997);
-
-  bool loading = true;
-  GoogleMapController? mapController;
-  LatLng? markerPosition = default_location;
-
-  Position? currentPosition;
-
-  @override
-  void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
-      setupMap();
-    });
-    super.initState();
-  }
-
-  void setupMap() async {
-    try {
-      //Check location service
-      if (!(await Geolocator.isLocationServiceEnabled())) {
-        showGetSnackBar('Location Service',
-            'Location services are not enabled on your device');
-        throw Exception();
-      }
-      //Check location permissions
-      final LocationPermission permission = await Geolocator.checkPermission();
-      if (permission == LocationPermission.denied) {
-        await Geolocator.requestPermission();
-      }
-      if (permission == LocationPermission.deniedForever) {
-        showGetSnackBar(
-            'Location Service', 'Location permissions are denied forever');
-      }
-      currentPosition = await Geolocator.getCurrentPosition();
-    } catch (_) {
-    } finally {
-      setState(() {
-        loading = false;
-      });
-    }
-  }
-
-  void completeMap() {
-    //Check if location is already selected
-    final arguments = Get.arguments;
-    if (arguments != null) {
-      final LatLng selectedLocation = Get.arguments as LatLng;
-      updateLocation(selectedLocation);
-    } else {
-      if (currentPosition != null) {
-        final LatLng latlng =
-            LatLng(currentPosition!.latitude, currentPosition!.longitude);
-        updateLocation(latlng);
-      } else {
-        updateLocation(default_location);
-      }
-    }
-  }
-
-  void updateLocation(LatLng latlng) {
-    markerPosition = latlng;
-    //Animate map
-    mapController!.animateCamera(
-      CameraUpdate.newCameraPosition(
-        CameraPosition(target: latlng, zoom: 15.0),
-      ),
-    );
-  }
 
   Widget confirmButton() {
     return Positioned(
@@ -95,32 +20,97 @@ class _MapViewState extends State<MapView> {
             backgroundColor: ColorConstants.accentColor),
         onPressed: () {
           Get.back(result: {
-            'lat': markerPosition!.latitude,
-            'lng': markerPosition!.longitude
+            'lat': controller.state.markerPosition.value.latitude,
+            'lng': controller.state.markerPosition.value.longitude
           });
         },
-        child: const Text('Confirm'),
+        child: const Text(
+          'Confirm',
+          style: TextStyle(color: Colors.white),
+        ),
       ),
     );
   }
 
   Widget backButton() {
+    return GestureDetector(
+      onTap: () {
+        Get.back();
+      },
+      child: Container(
+        padding: EdgeInsets.all(8.w),
+        decoration: const BoxDecoration(
+            shape: BoxShape.circle, color: ColorConstants.primaryColor),
+        child: Icon(
+          Icons.arrow_back,
+          color: Colors.white,
+          size: 20.w,
+        ),
+      ),
+    );
+  }
+
+  Widget currentLocationIcon() {
+    return GestureDetector(
+      onTap: () {
+        controller.updateLocation(LatLng(
+            controller.state.currentPosition!.latitude,
+            controller.state.currentPosition!.longitude));
+      },
+      child: Container(
+        padding: EdgeInsets.all(8.w),
+        decoration: const BoxDecoration(
+            shape: BoxShape.circle, color: ColorConstants.primaryColor),
+        child: Icon(
+          Icons.my_location_sharp,
+          color: Colors.white,
+          size: 20.w,
+        ),
+      ),
+    );
+  }
+
+  Widget searchPlaces() {
+    return Container(
+      margin: EdgeInsets.symmetric(horizontal: 10.w),
+      decoration: BoxDecoration(
+          color: ColorConstants.primaryColor,
+          borderRadius: BorderRadius.circular(40.h)),
+      width: double.infinity,
+      child: ListTile(
+        onTap: () {
+          controller.searchPlace();
+        },
+        leading: const Icon(
+          Icons.search,
+          color: Colors.white,
+        ),
+        title: Text(
+          controller.state.searchLocationName.value ?? 'Search a location',
+          maxLines: 1,
+          overflow: TextOverflow.ellipsis,
+          style: const TextStyle(color: Colors.white),
+        ),
+      ),
+    );
+  }
+
+  Widget topArea() {
     return Positioned(
       top: 50.h,
-      left: 13.w,
-      child: GestureDetector(
-        onTap: () {
-          Get.back();
-        },
-        child: Container(
-          padding: EdgeInsets.all(8.w),
-          decoration: const BoxDecoration(
-              shape: BoxShape.circle, color: ColorConstants.primaryColor),
-          child: Icon(
-            Icons.arrow_back,
-            color: Colors.white,
-            size: 20.w,
-          ),
+      left: 0,
+      right: 0,
+      child: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 14.w),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            backButton(),
+            Expanded(
+              child: searchPlaces(),
+            ),
+            currentLocationIcon(),
+          ],
         ),
       ),
     );
@@ -129,40 +119,42 @@ class _MapViewState extends State<MapView> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: loading
-          ? const Center(
-              child: LoadingIndicator(),
-            )
-          : Stack(
-              children: [
-                GoogleMap(
-                  markers: markerPosition == null
-                      ? {}
-                      : {
-                          Marker(
-                            markerId: const MarkerId('selectedLocation'),
-                            position: markerPosition!,
-                          ),
-                        },
-                  onCameraMove: (CameraPosition position) {
-                    setState(() {
-                      markerPosition = position.target;
-                    });
-                  },
-                  onMapCreated: (GoogleMapController controller) {
-                    mapController = controller;
-                    completeMap();
-                  },
-                  initialCameraPosition: const CameraPosition(
-                    target: default_location,
-                    zoom: 12.0,
-                  ),
-                  myLocationEnabled: true,
+      body: Obx(() {
+        if (controller.state.loading.isTrue) {
+          return const Center(
+            child: LoadingIndicator(),
+          );
+        }
+        return Stack(
+          children: [
+            GoogleMap(
+              zoomControlsEnabled: false,
+              zoomGesturesEnabled: true,
+              markers: {
+                Marker(
+                  markerId: const MarkerId('selectedLocation'),
+                  position: controller.state.markerPosition.value,
                 ),
-                confirmButton(),
-                backButton(),
-              ],
+              },
+              onCameraMove: (CameraPosition position) {
+                controller.state.markerPosition.value = position.target;
+              },
+              onMapCreated: (GoogleMapController controller) {
+                this.controller.state.mapController = controller;
+                this.controller.completeMap();
+              },
+              initialCameraPosition: const CameraPosition(
+                target: MapState.default_location,
+                zoom: 12.0,
+              ),
+              myLocationEnabled: true,
+              myLocationButtonEnabled: false,
             ),
+            confirmButton(),
+            topArea(),
+          ],
+        );
+      }),
     );
   }
 }
