@@ -7,20 +7,21 @@ import 'constants.dart';
 
 //Validations
 bool isValidEmail(String email) {
-  RegExp emailRegex = RegExp(
-      r'^[\w-]+(\.[\w-]+)*@[A-Za-z0-9]+(\.[A-Za-z0-9]+)*(\.[A-Za-z]{2,})$');
+  final RegExp emailRegex =
+      RegExp(r'^[\w-]+(\.[\w-]+)*@[A-Za-z0-9]+(\.[A-Za-z]{2,})$');
+
   return emailRegex.hasMatch(email);
 }
 
 bool isValidPassword(String password) {
-  RegExp passwordRegex = RegExp(
+  final RegExp passwordRegex = RegExp(
       r'^(?=.*[A-Z])(?=.*[!@#$%^&*(),.?":{}|<>])[A-Za-z\d!@#$%^&*(),.?":{}|<>]{8,}$');
 
   return passwordRegex.hasMatch(password);
 }
 
 bool isValidFullName(String fullName) {
-  RegExp fullNameRegex = RegExp(r"^[A-Za-z\s\'\-]+$");
+  final RegExp fullNameRegex = RegExp(r"^[A-Za-z\s\'\-]+$");
   return fullNameRegex.hasMatch(fullName);
 }
 
